// apps/frontend/src/api/generated/model/quizDetailDtoSlidesItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * QuizDock API
 * API REST du builder de quiz et des restitutions
 * OpenAPI spec version: 0.1.0
 */
import type { QuizDetailDtoSlidesItemAudioTarget } from './quizDetailDtoSlidesItemAudioTarget';
import type { QuizDetailDtoSlidesItemBlocksItem } from './quizDetailDtoSlidesItemBlocksItem';
import type { QuizDetailDtoSlidesItemGradient } from './quizDetailDtoSlidesItemGradient';
import type { QuizDetailDtoSlidesItemTextTone } from './quizDetailDtoSlidesItemTextTone';
import type { QuizDetailDtoSlidesItemWaveformSize } from './quizDetailDtoSlidesItemWaveformSize';

export type QuizDetailDtoSlidesItem = {
  id: string;
  quizId: string;
  /** @nullable */
  beforeQuestionId: string | null;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  orderIndex: number;
  blocks: QuizDetailDtoSlidesItemBlocksItem[];
  /** @nullable */
  mediaId: string | null;
  /** @nullable */
  gradient: QuizDetailDtoSlidesItemGradient;
  /** @nullable */
  videoMediaId: string | null;
  videoLoop: boolean;
  videoSound: boolean;
  /** @nullable */
  audioMediaId: string | null;
  waveformSize: QuizDetailDtoSlidesItemWaveformSize;
  /** @nullable */
  audioTarget: QuizDetailDtoSlidesItemAudioTarget;
  textTone: QuizDetailDtoSlidesItemTextTone;
  textOutline: boolean;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     * @nullable
     */
  displayDelayS: number | null;
};